Move SpecmaticConfigTest to core package
